requirements.txt
asgiref==3.8.1
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
sqlparse==0.5.0
djangorestframework-simplejwt==5.3.1